Refactor StockTicker schema and API imports
- Move StockTickerOut schema to a new backend/schemas.py file
- Remove unused imports from api.py and core middleware/models
- Clean up migration files formatting
