+++ backend/core/models.py
@@ -1,5 +1,4 @@
 import uuid
-from decimal import Decimal
 
 from django.db import models
 from django.contrib.auth.models import AbstractUser, Group, Permission
